requirements.txt
aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.8.0
argparse==1.4.0
attrs==25.1.0
babel==2.17.0
backoff==2.2.1
bashlex==0.18
beautifulsoup4==4.13.3
boto3==1.37.6
botocore==1.37.6
browser-use==0.1.36
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
courlan==1.3.2
dateparser==1.2.1
python-dateutil==2.9.0.post0
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
fastapi==0.115.11
filetype==1.2.0
fireworks-ai==0.15.12
frozenlist==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.162.0
googleapis-common-protos==1.69.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
greenlet==3.1.1
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
html2text==2024.2.26
htmldate==1.9.3
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.1
idna==3.10
importlib-metadata==8.5.0
jiter==0.8.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
justext==3.0.2
langchain==0.3.14
langchain_anthropic==0.3.3
langchain_aws==0.2.15
langchain_core==0.3.41
langchain_fireworks==0.2.7
langchain_google_genai==2.0.8
langchain_ollama==0.2.2
langchain_openai==0.3.1
langchain_text_splitters==0.3.6
langsmith==0.2.11
lmnr==0.4.62
lxml==5.3.1
MainContentExtractor==0.0.4
markdownify==0.14.1
monotonic==1.6
multidict==6.1.0
numpy==1.26.4
ollama==0.4.7
openai==1.65.3
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-langchain==0.38.10
opentelemetry-instrumentation-requests==0.51b0
opentelemetry-instrumentation-sqlalchemy==0.51b0
opentelemetry-instrumentation-threading==0.51b0
opentelemetry-instrumentation-urllib3==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-semantic-conventions-ai==0.4.3
opentelemetry-util-http==0.51b0
orjson==3.10.15
packaging==24.2
pexpect==4.9.0
pillow==11.1.0
playwright==1.50.0
posthog==3.18.1
propcache==0.3.0
protobuf==5.29.3
proto-plus==1.26.0
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.10.6
pydantic-core==2.27.2
pyee==12.1.1
pyparsing==3.2.1
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.46.0
tenacity==9.0.0
tiktoken==0.9.0
tld==0.13
tqdm==4.67.1
trafilatura==2.0.0
typing-extensions==4.12.2
tzlocal==5.3
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0